code/requirements.txt
langchain-openai>=0.1.0
langchain>=0.2.0
langchain-community>=0.2.0
langgraph>=0.1.0
langchain-tavily>=0.1.0
faiss-cpu>=1.7.4
openai>=1.30.0
streamlit>=1.35.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tiktoken>=0.7.0
